# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API сервис для платформы копитрейдинга",
        "title": "Copy Trading API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/statistics/leaderboard": {
            "get": {
                "description": "Возвращает топ стратегий по доходности",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "statistics"
                ],
                "summary": "Лидерборд стратегий",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Количество записей",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/statistics.LeaderboardEntry"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/statistics/master-income": {
            "get": {
                "description": "Возвращает информацию о доходах мастер-трейдера",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "statistics"
                ],
                "summary": "Доход мастера",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID пользователя-мастера",
                        "name": "user_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Начало периода (RFC3339)",
                        "name": "from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Конец периода (RFC3339)",
                        "name": "to",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/statistics.MasterIncome"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/statistics/portfolio": {
            "get": {
                "description": "Возвращает портфель инвестора с его подписками и статистикой",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "statistics"
                ],
                "summary": "Портфель инвестора",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID пользователя-инвестора",
                        "name": "user_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/statistics.PortfolioEntry"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "statistics.LeaderboardEntry": {
            "type": "object",
            "properties": {
                "active_subscriptions": {
                    "type": "integer"
                },
                "strategy_id": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "total_commissions": {
                    "type": "number"
                },
                "total_profit": {
                    "type": "number"
                }
            }
        },
        "statistics.MasterIncome": {
            "type": "object",
            "properties": {
                "management_fees": {
                    "type": "number"
                },
                "performance_fees": {
                    "type": "number"
                },
                "registration_fees": {
                    "type": "number"
                },
                "total_income": {
                    "type": "number"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "statistics.PortfolioEntry": {
            "type": "object",
            "properties": {
                "copied_trades_count": {
                    "type": "integer"
                },
                "strategy_id": {
                    "type": "integer"
                },
                "strategy_title": {
                    "type": "string"
                },
                "subscription_id": {
                    "type": "integer"
                },
                "total_profit": {
                    "type": "number"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}